fix: low-priority correctness/robustness/perf polish from #48
Batch cleanup of the discrete items in #48. Larger directional items
(object-store structured error codes, PageCache Arc bodies, sealed_bytes
memoization) are deferred to follow-up issues.

Dead/misleading logic:
- vtl verify CLI: collapse the dead three-way log branch (all arms ran
  the same eprintln!) into a one-liner, matching the VSA equivalent.
- vtl daemon: remove dead run_upload_worker (the daemon spawns
  run_event_driven_upload_worker; this had no caller).

Robustness / silent failure:
- vtl library restore CLI: flow cli_result through record_result and
  derive the exit code from its return instead of discarding it via
  `let _ =`.
- vsa volume destroy CLI: distinguish a daemon-omitted `lun` field from
  a real LUN 0 (print without the LUN when absent).
- alerting AlertRateLimiter: drain expired windows on each allow() so
  fire-once dedup keys don't accumulate forever (no flush task here,
  unlike the audit-side limiter).

Overflow / truncation:
- dirty_pages mark_range: do the page-index division in u64 and
  debug_assert the u32 bound (~256 TiB) instead of an unguarded narrow.
- ssc WRITE ATTRIBUTE: compare lengths as u64 so a host-controlled
  `payload_len + 4` can't wrap on a 32-bit usize and pass the check.
- pool budget try_reserve: saturating_add to match add/force_reserve.

Concurrency hardening:
- chunk_index / block_index: document the single-writer invariant at
  the type (the AtomicU64 next_id is interior mutability behind &self,
  not concurrent-append safety; the owning Cartridge serializes via
  &mut self).
- iscsi CHAP verify: constant-time response comparison via
  subtle::ConstantTimeEq (hardening; per-login challenge keeps risk low).

Micro-perf:
- nvme WRITE ZEROES: chunk the zero buffer at 16 MiB like SBC WRITE SAME
  instead of materializing the full transfer length.
- iscsi Data-In: add write_pdu_with_data so each chunk writes straight
  out of the owned response buffer, dropping the per-chunk to_vec clone.

+++ scsi/ssc/src/scsi/attributes.rs
@@ -193,10 +193,12 @@ pub fn handle_write_attribute(data: &[u8]) -> Result<(), String> {
         ));
     }
     let payload_len = u32::from_be_bytes([data[0], data[1], data[2], data[3]]) as usize;
-    if payload_len + 4 > data.len() {
+    // Compare as u64: `payload_len` is host-controlled and `+ 4` would
+    // wrap on a 32-bit `usize`, silently passing this bounds check.
+    if payload_len as u64 + 4 > data.len() as u64 {
         return Err(format!(
             "WRITE ATTRIBUTE: parameter list claims {} bytes but only {} available",
-            payload_len + 4,
+            payload_len as u64 + 4,
             data.len()
         ));
     }
@@ -207,7 +209,7 @@ pub fn handle_write_attribute(data: &[u8]) -> Result<(), String> {
         let attr_id = u16::from_be_bytes([data[offset], data[offset + 1]]);
         let _format = data[offset + 2];
         let attr_len = u16::from_be_bytes([data[offset + 4], data[offset + 5]]) as usize;
-        if offset + 6 + attr_len > end {
+        if offset as u64 + 6 + attr_len as u64 > end as u64 {
             return Err(format!(
                 "WRITE ATTRIBUTE: attribute 0x{:04x} declared length {} overflows parameter list",
                 attr_id, attr_len
